+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - **[MODULE]** Checksums (`binary/checksums`) _[from @HenryHoggard]_
 #### Fixed
 - **[CORE]** App metadata: creation of binary path from MobileInstallation.plist
+- **[CORE]** Plist wrapper using biplist
 - **[MODULE]** Frida Script: print view hierarchy (`hooking/frida/script_dump-ui`) _[from @HenryHoggard]_
 - **[MODULE]** Improved SQLite DB identification by reducing false positives and false negatives _[from @HenryHoggard]_
 #### Removed